app/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
streamlit
plotly
pandas